Move GlobalFilter files
