+++ src/components/global-filter/global-filter-autocomplete.tsx
@@ -25,11 +25,11 @@ import { FilterAlt, WarningAmberRounded } from '@mui/icons-material';
 import { useIntl } from 'react-intl';
 import { useLocalizedCountries } from 'components/utils/localized-countries-hook';
 import { useSelector } from 'react-redux';
-import { AppState } from '../../../../redux/reducer';
-import { FilterType } from '../utils';
+import { AppState } from '../../redux/reducer';
+import { FilterType } from '../results/common/utils';
 import { OverflowableText } from '@gridsuite/commons-ui';
-import { EQUIPMENT_TYPES } from '../../../utils/equipment-types';
-import { GlobalFilter } from './global-filter-types';
+import { EQUIPMENT_TYPES } from '../utils/equipment-types';
+import type { GlobalFilter } from './types';
 import { getResultsGlobalFiltersChipStyle, resultsGlobalFilterStyles } from './global-filter-styles';
 import GlobalFilterPaper from './global-filter-paper';
 import Tooltip from '@mui/material/Tooltip';
