+++ src/components/global-filter/global-filter-provider.tsx
@@ -6,16 +6,16 @@
  */
 
 import { PropsWithChildren, useCallback, useEffect, useMemo, useState } from 'react';
-import { GlobalFilter } from './global-filter-types';
-import { FilterType } from '../utils';
+import { FilterType } from '../results/common/utils';
 import { UUID } from 'crypto';
 import { ElementAttributes, fetchDirectoryElementPath, useSnackMessage } from '@gridsuite/commons-ui';
-import { computeFullPath } from '../../../../utils/compute-title';
-import { addToRecentGlobalFilters, removeFromRecentGlobalFilters } from '../../../../redux/actions';
+import type { GlobalFilter } from './types';
+import { computeFullPath } from '../../utils/compute-title';
+import { addToRecentGlobalFilters, removeFromRecentGlobalFilters } from '../../redux/actions';
 import { useDispatch } from 'react-redux';
-import { AppDispatch } from '../../../../redux/store';
+import { AppDispatch } from '../../redux/store';
 import { GlobalFilterContext } from './global-filter-context';
-import { HttpStatusCode } from '../../../../utils/http-status-code';
+import { HttpStatusCode } from '../../utils/http-status-code';
 
 export default function GlobalFilterProvider({
     children,
