+++ src/components/results/common/global-filter/use-global-filters.ts
@@ -6,7 +6,8 @@
  */
 
 import { useCallback, useState } from 'react';
-import { GlobalFilter, GlobalFilters } from './global-filter-types';
+import type { GlobalFilter } from '../../../global-filter/types';
+import { GlobalFilters } from '../types';
 import { FilterType } from '../utils';
 
 interface UseGlobalFiltersParams {
