+++ src/components/results/shortcircuit/shortcircuit-analysis-result.tsx
@@ -38,7 +38,7 @@ import { FilterType, PaginationType, ShortcircuitAnalysisTab } from '../../../ty
 import { mapFieldsToColumnsFilter } from '../../../utils/aggrid-headers-utils';
 import { FilterEnumsType } from '../../custom-aggrid/custom-aggrid-filters/custom-aggrid-filter.type';
 import { usePaginationSelector } from 'hooks/use-pagination-selector';
-import { GlobalFilters } from '../common/global-filter/global-filter-types';
+import { GlobalFilters } from '../common/types';
 
 interface IShortCircuitAnalysisGlobalResultProps {
     analysisType: ShortCircuitAnalysisType;
