+++ src/services/study/sensitivity-analysis.ts
@@ -17,7 +17,7 @@ import {
     SensitivityResultFilterOptions,
 } from './sensitivity-analysis.type';
 import { FilterConfig } from '../../types/custom-aggrid-types';
-import { GlobalFilters } from 'components/results/common/global-filter/global-filter-types';
+import { GlobalFilters } from 'components/results/common/types';
 
 const GET_PARAMETERS_PREFIX = import.meta.env.VITE_API_GATEWAY + '/sensitivity-analysis/v1/parameters';
