+++ src/services/study/voltage-init.ts
@@ -9,7 +9,7 @@ import { getStudyUrl, getStudyUrlWithNodeUuidAndRootNetworkUuid } from './index'
 import { backendFetch, backendFetchJson, backendFetchText } from '../utils';
 import { UUID } from 'crypto';
 import { VoltageInitStudyParameters } from '@gridsuite/commons-ui';
-import { ResultsQueryParams } from '../../components/results/common/global-filter/global-filter-types';
+import { ResultsQueryParams } from '../../components/results/common/types';
 
 export function startVoltageInit(
     studyUuid: UUID,
